updated models, factory for Employee and Task. Local performance test for cache: For 1000 employees and 1000 tasks, the request takes 12.63 seconds using a cache for hours. For 1000 employees and 1000 tasks, the request takes 16.91 seconds without using a cache for hours.

+++ TalaTask/tasks/models.py
@@ -19,9 +19,8 @@ class Employee(models.Model):
     name = models.CharField(max_length=100)
     skills = models.ManyToManyField('Skill')   
     #this is average hours per day. For example, if the Employee has 2 days available, and 10 hours, it means that
-    #the user has 20 available hours.
-    #so, if the user takes a 10 hours task, the field available_hours_per_day take the value of 5.    
-    available_hours_per_day = models.DecimalField(max_digits=5, decimal_places=2)
+    #the user has 20 available hours.  
+    available_hours_per_day = models.IntegerField()
     available_days = MultiSelectField(choices=DAY_CHOICES)
 
     def __str__(self):
